requirements.txt
# Framework
pytest~=6.1.2
selenium~=3.141.0
pycmd

# Reporting
allure-pytest

# Data Driven
pyexcel~=0.6.7
pyexcel-xls
pyexcel-xlsx

# Driver Manager
webdriver_manager
py~=1.10.0
colorama~=0.4.3
lml~=0.1.0
pip~=20.0.2
configparser~=5.0.2
wheel~=0.34.2
cryptography~=2.8
lxml~=4.6.3
keyring~=18.0.1
attrs~=21.2.0
toml~=0.10.2
wcwidth~=0.2.5
pluggy~=0.13.1
iniconfig~=1.1.1
pathlib~=1.0.1
pexpect~=4.6.0
certifi~=2019.11.28
chardet~=3.0.4
texttable~=1.6.4
urllib3~=1.26.8
openpyxl~=3.0.9
defusedxml~=0.7.1
pandas~=1.3.3
numpy~=1.21.2
Pillow~=8.1.2
idna~=2.8
simplejson~=3.16.0
pyparsing~=2.4.7
six~=1.16.0
setuptools~=45.2.0
future~=0.18.2
xlrd~=2.0.1
xlwt~=1.3.0
requests~=2.25.1
pjsua~=2.11.1